--- app/api/chat/route.ts
@@ -1,11 +1,0 @@
-import { mastra } from "@/mastra";
-import { NextResponse } from "next/server";
-
-export async function POST(req: Request) {
-  const { city } = await req.json();
-  const agent = mastra.getAgent("weatherAgent");
-
-  const result = await agent.stream(`What's the weather like in ${city}?`);
-
-  return result.toDataStreamResponse();
-}
